(hel-668) Renomage du composant Etablissement en TuileEtablissement
